PC - array indexing must be integer

+++ src/Components/missions/PACE/mcd43c.py
@@ -243,12 +243,12 @@ class MCD43C(object):
                     lon[i] = lon[i] + 360.0
 
                     nlat,nlon = data.shape
-                    lon1 = self.lon[0:nlon*0.5]
-                    lon2 = self.lon[nlon*0.5:]
-                    datalon = np.append(lon2,lon1+360.)
+                    lon1 = self.lon[0:int(nlon*0.5)]
+                    lon2 = self.lon[int(nlon*0.5):]
+                    datalon = np.append(lon2,lon1+360)
 
-                    data1 = data[:,0:nlon*0.5]
-                    data2 = data[:,nlon*0.5:]
+                    data1 = data[:,0:int(nlon*0.5)]
+                    data2 = data[:,int(nlon*0.5):]
                     data  = np.ma.hstack((data2, data1))
                 else:
                     datalon = self.lon
